Add test for try dispatch

+++ modules/src/ics03_connection/context.rs
@@ -17,7 +17,7 @@ pub trait ConnectionReader {
     fn chain_current_height(&self) -> Height;
 
     /// Returns the number of consensus state historical entries for the local chain.
-    fn chain_consensus_states_history_size(&self) -> u32;
+    fn chain_consensus_states_history_size(&self) -> usize;
 
     /// Returns the prefix that the local chain uses in the KV store.
     fn commitment_prefix(&self) -> CommitmentPrefix;
